Only param row is refactored
- Only param row is refactored (originally all list items were)
- Param row has all specified props
- All custom component definitions moved to their own files
- Add Block list tile component removed
- DrawerHeader gets boxShadow to match main's toolbar
